refactor: Move global/ to pkg/version
