+++ cmd/commands/version_test.go
@@ -4,8 +4,8 @@ import (
 	"testing"
 
 	"github.com/caffeine-addictt/waku/cmd/commands"
-	"github.com/caffeine-addictt/waku/cmd/global"
 	"github.com/caffeine-addictt/waku/cmd/helpers"
+	"github.com/caffeine-addictt/waku/pkg/version"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -13,6 +13,6 @@ func TestVersionOut(t *testing.T) {
 	stdout, stderr, err := helpers.ExecuteCommand(commands.VersionCmd, []string{})
 	assert.NoError(t, err)
 
-	assert.Equal(t, stdout, global.Version+"\n")
+	assert.Equal(t, stdout, version.Version+"\n")
 	assert.Empty(t, stderr)
 }
