+++ pkg/version/version.go
@@ -1,4 +1,4 @@
-package global
+package version
 
 // The current app version
 const Version = "0.3.3"
