+++ pkg/version/version_test.go
@@ -1,10 +1,10 @@
-package global_test
+package version_test
 
 import (
 	"regexp"
 	"testing"
 
-	"github.com/caffeine-addictt/waku/cmd/global"
+	"github.com/caffeine-addictt/waku/pkg/version"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -12,5 +12,5 @@ import (
 var semverRegex = regexp.MustCompile(`^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)(?:-((?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\.(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\+([0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?$`)
 
 func TestFollowsSemVer(t *testing.T) {
-	assert.True(t, semverRegex.MatchString(global.Version), "%s does not follow semver", global.Version)
+	assert.True(t, semverRegex.MatchString(version.Version), "%s does not follow semver", version.Version)
 }
